Add 'cache' mode
This mode records HTTP requests that don't already have associated
fixtures, but then plays back fixtures that already exist. This will
allow accelerating testing without making it very fragile and sensitive
to small changes in the requests we make.

Currently, the code for the cache mode has a lot of overlap from both
the record and playback modes. This is intentional, because we first
want to verify, through repeated usage, whether this functionality works
as intended. After ironing out any bugs, we can start refactoring the
code.