Create .missing files in playback
When not recording missing fixtures, create a <fixture-name>.missing
file containing information about the request for which a fixture could
not be found. This is in addition to throwing an error, as was the case
previously.

This allows the creation of tools that either watch for the creation of
these .missing files, or apply some post-analysis on these files, in
order to report what failures occurred when attempting to access
fixtures.