Spin up a small HTTP server to set test options
Sepia now starts a small embedded HTTP server when the module is
required. This server runs in the same process in which sepia runs, so
from an external source, the server can be hit in order to modify
in-process settings.

In this case, a "test name" is configurable via the server, in order to
emulate a type of "cassette" functionality available in other VCR
implementations, allowing fixtures to be separated into various
directories.

Currently, the server's port is hardcoded, but this will be configurable
in the future.