Second round of auditing and cleaning up
Refactored major sections of util.js, adding tests for almost all the
functionality (everything except the straightforward "setters").

The only API change here is the removal of "addRecordOverride". This
functionality is folded in an option ("forceLive") in the filters,
because it operates like a filter based on a regex match on the URL.

Also added a small example file for the "forceLive" functionality. In
later commits, I'll break up the examples into smaller files,
refactoring some of the code and trying to make some of the checks more
automated so it becomes more like a test suite.