Broke up examples by feature
In doing so, I abstracted out common functionality used by the examples,
and introduced 'should' into the mix. With this, the examples also
become automated tests that can be used to determine the correctness of
the system.

This cleanup resulted in all the examples passing jshint, so I enabled
that in 'npm test'.

The README.md is now outdated with respect to the examples, so that will
have to be updated when I re-do the documentation.