Fix #8: don't start up server by default
The embedded server does not really need to be started if its
functionality is not required. Why waste resources on another server
after all?

It turns out (and I should have thought through this before) that the
process.exit() calls in the examples were required precisely because
this server didn't shut down. I added a sepia.shutdown() function to
address this.

I bumped the major version because this is technically a
backwards-incompatible change.