Fix #10: Support for node 0.10.x-style APIs
When sepia was tested with the then-request module, it turned out that
sepia was not compatible with that module due not conforming to certain
updated node behavior in the standard library.

Of course, normally, this type of behavior is transparent to users of
node, but sepia has to emulate that behavior, hence the breakage.

Unfortunately, the updated behavior is not compatible with 0.8.x, which
is an existing use case. I'll deprecate it eventually, but not yet. To
support both 0.8.x and 0.10.x, I added some conditional code.