Use correct timestamp for fs.utimesSync
As per the node.js documentation for `fs.utimes`:

> The value should be a Unix timestamp in seconds. For example,
> `Date.now()` returns milliseconds, so it should be divided by 1000
> before passing it in.

I'm not sure why this wasn't causing a problem for most people, but it
was reported by @mseminatore as a problem on Windows, on node.js
v0.12.2. In any case, this fix doesn't cause a problem on Linux, on
node.js v6.9.1, so I'm happy to make the fix.